sdk/js: Moved consts and helpers under utils directory
