+++ sdk/js/src/nft_bridge/__tests__/utils/consts.ts
