+++ sdk/js/src/token_bridge/__tests__/solana-integration.ts
@@ -41,7 +41,7 @@ import {
   SOLANA_PRIVATE_KEY,
   TEST_SOLANA_TOKEN,
   WORMHOLE_RPC_HOSTS,
-} from "./consts";
+} from "./utils/consts";
 
 jest.setTimeout(60000);
